requirements/development.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
Deprecated==1.2.14
Django==4.2.4
django-autoslug==1.9.9
django-axes==5.41.1
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-dirtyfields==1.9.2
django-dotenv==1.4.2
django-filter==23.1
django-ipware==5.0.0
django-phonenumber-field==7.0.2
django-redis==5.3.0
django-silk==5.0.3
django-simple-history==3.3.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
et-xmlfile==1.1.0
Faker==18.4.0
gprof2dot==2022.7.29
hiredis==2.2.3
idna==3.4
inflection==0.5.1
infobip-api-python-sdk==5.0.1
jsonschema==4.17.3
Markdown==3.4.3
more-itertools==9.1.0
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
phonenumbers==8.13.10
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-ulid==1.1.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
redis-om==0.2.1
regex==2022.3.15
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.3
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wrapt==1.15.0
